refactor: 인게임 소켓 로직 통합
